Moved all the logic code from report subcomponents to the report component
