+++ src/app/components/report/burned-areas-chart/burned-areas-chart.component.html
@@ -1,13 +1,14 @@
 <div class="p-grid p-nogutter">
-    <div class="p-col-12 p-col-nogutter mb-5">
+    <div class="p-col-12 p-col-nogutter">
+        <p-chart type="bar" [data]="data1"></p-chart>
     </div>
-    <div class="p-col-12 p-col-nogutter p-col-md-4 p-0">
-        <div id="area-pie-chart1" class="chart" style="position: relative;"></div>
+    <div class="p-col-12 p-col-nogutter p-col-md-4">
+        <p-chart type="pie" [data]="data2"></p-chart>
     </div>
-    <div class="p-col-12 p-col-nogutter p-col-md-4 p-0">
-        <div id="area-pie-chart2" class="chart" style="position: relative;"></div>
+    <div class="p-col-12 p-col-nogutter p-col-md-4">
+        <p-chart type="pie" [data]="data3"></p-chart>
     </div>
-    <div class="p-col-12 p-col-nogutter p-col-md-4 p-0">
-        <div id="area-pie-chart3" class="chart" style="position: relative;"></div>
+    <div class="p-col-12 p-col-nogutter p-col-md-4">
+        <p-chart type="pie" [data]="data4"></p-chart>
     </div>
 </div>
