+++ src/app/components/report/burned-areas-chart/burned-areas-chart.component.ts
@@ -1,15 +1,17 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-burned-areas-chart',
   templateUrl: './burned-areas-chart.component.html',
   styleUrls: ['./burned-areas-chart.component.css']
 })
-export class BurnedAreasChartComponent implements OnInit {
+export class BurnedAreasChartComponent {
+
+  data1;
+  data2;
+  data3;
+  data4;
 
   constructor() { }
-
-  ngOnInit() {
-  }
 
 }
