+++ src/app/components/report/burning-focus-chart/burning-focus-chart.component.html
@@ -1,0 +1,1 @@
+<p-chart type="bar" [data]="data"></p-chart>
