+++ src/app/components/report/deforestation/deforestation.component.ts
@@ -1,40 +1,16 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 import { Vision } from 'src/app/models/vision.model';
-
-import { Property } from 'src/app/models/property.model';
-
-import { ConfigService } from 'src/app/services/config.service';
 
 @Component({
   selector: 'app-deforestation',
   templateUrl: './deforestation.component.html',
   styleUrls: ['./deforestation.component.css']
 })
-export class DeforestationComponent implements OnInit {
+export class DeforestationComponent {
 
-  @Input() property: Property;
+  @Input() deforestations: Vision[] = [];
 
-  deforestations: Vision[] = [];
-
-  constructor(
-    private configService: ConfigService
-  ) { }
-
-  ngOnInit() {
-    const deforestationsData = this.configService.getConfig('report').deforestations;
-
-    deforestationsData.forEach((deforestationData: Vision) => {
-      const image = deforestationData.image.replace('{bbox}', this.property.bbox);
-      const vision = new Vision(
-        deforestationData.id,
-        deforestationData.title,
-        image,
-        deforestationData.description,
-        deforestationData.layerData
-      );
-      this.deforestations.push(vision);
-    });
-  }
+  constructor() { }
 
 }
