+++ src/app/components/report/image-history/image-history.component.ts
@@ -1,40 +1,16 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 import { Vision } from 'src/app/models/vision.model';
-
-import { Property } from 'src/app/models/property.model';
-
-import { ConfigService } from 'src/app/services/config.service';
 
 @Component({
   selector: 'app-image-history',
   templateUrl: './image-history.component.html',
   styleUrls: ['./image-history.component.css']
 })
-export class ImageHistoryComponent implements OnInit {
+export class ImageHistoryComponent {
 
-  @Input() property: Property;
+  @Input() landsatHistories: Vision[] = [];
 
-  landsatHistories: Vision[] = [];
-
-  constructor(
-    private configService: ConfigService
-  ) { }
-
-  ngOnInit() {
-    const landsatHistoriesData = this.configService.getConfig('report').landsatHistories;
-
-    landsatHistoriesData.forEach((landsatHistoryData: Vision) => {
-      const image = landsatHistoryData.image.replace('{bbox}', this.property.bbox);
-      const vision = new Vision(
-        landsatHistoryData.id,
-        landsatHistoryData.title,
-        image,
-        landsatHistoryData.description,
-        landsatHistoryData.layerData
-      );
-      this.landsatHistories.push(vision);
-    });
-  }
+  constructor() { }
 
 }
