+++ src/app/components/report/report.component.html
@@ -5,22 +5,22 @@
       <!-- <h1>Relatório</h1> -->
       <app-property-data [property]="property"></app-property-data>
       <hr>
-      <app-vision [visions]="visions" [legends]="legends"></app-vision>
+      <app-vision [visions]="visions" [legends]="visionLegends"></app-vision>
       <br>
-      <app-vision-detailed [property]="property"></app-vision-detailed>
+      <app-vision-detailed [detailedVisions]="detailedVisions"></app-vision-detailed>
       <br>
-      <app-deforestation [property]="property"></app-deforestation>
+      <app-deforestation [deforestations]="deforestations"></app-deforestation>
       <br>
       <br>
-      <app-deforestation-history [property]="property"></app-deforestation-history>
+      <app-deforestation-history [deforestationHistories]="deforestationHistories"></app-deforestation-history>
       <br>
-      <app-burning-spotlights [property]="property"></app-burning-spotlights>
+      <app-burning-spotlights [burningSpotlights]="burningSpotlights"></app-burning-spotlights>
       <br>
       <br>
       <app-burning-focus-chart></app-burning-focus-chart>
       <br>
       <br>
-      <app-burned-areas [property]="property"></app-burned-areas>
+      <app-burned-areas [burnedAreas]="burnedAreas"></app-burned-areas>
       <br>
       <app-burned-areas-chart></app-burned-areas-chart>
       <br>
