+++ src/app/components/report/report.component.ts
@@ -27,7 +27,19 @@ export class ReportComponent implements OnInit {
 
   visions: Vision[] = [];
 
-  legends: Legend[] = [];
+  detailedVisions: Vision[] = [];
+
+  deforestations: Vision[] = [];
+
+  deforestationHistories: Vision[] = [];
+
+  burningSpotlights: Vision[] = [];
+
+  burnedAreas: Vision[] = [];
+
+  landsatHistories: Vision[] = [];
+
+  visionLegends: Legend[] = [];
 
   constructor(
     private activatedRoute: ActivatedRoute,
@@ -38,6 +50,7 @@ export class ReportComponent implements OnInit {
   ngOnInit() {
     this.activatedRoute.params.subscribe(params => this.carRegister = params.carRegister);
     this.reportConfig = this.configService.getConfig('report');
+    this.visionLegends = this.reportConfig.visionslegends;
     this.getPropertyData();
   }
 
@@ -58,7 +71,7 @@ export class ReportComponent implements OnInit {
                                   propertyData['city'],
                                   bbox,
                                   latLong,
-                                  );
+      );
       const visionsData = this.reportConfig.visions;
       visionsData.forEach((visionData: Vision) => {
         const image = visionData.image.replace('{bbox}', this.property.bbox);
@@ -71,7 +84,84 @@ export class ReportComponent implements OnInit {
         );
         this.visions.push(vision);
       });
+
+      const detailedVisionsData = this.reportConfig.detailedVisions;
+      detailedVisionsData.forEach((detailedVisionData: Vision) => {
+        const image = detailedVisionData.image.replace('{bbox}', this.property.bbox);
+        const vision = new Vision(
+          detailedVisionData.id,
+          detailedVisionData.title,
+          image,
+          detailedVisionData.description,
+          detailedVisionData.layerData
+        );
+        this.detailedVisions.push(vision);
+      });
+
+      const deforestationsData = this.reportConfig.deforestations;
+      deforestationsData.forEach((deforestationData: Vision) => {
+        const image = deforestationData.image.replace('{bbox}', this.property.bbox);
+        const vision = new Vision(
+          deforestationData.id,
+          deforestationData.title,
+          image,
+          deforestationData.description,
+          deforestationData.layerData
+        );
+        this.deforestations.push(vision);
+      });
+
+      const deforestationHistoriesData = this.reportConfig.deforestationHistories;
+      deforestationHistoriesData.forEach((deforestationHistoryData: Vision) => {
+        const image = deforestationHistoryData.image.replace('{bbox}', this.property.bbox);
+        const vision = new Vision(
+          deforestationHistoryData.id,
+          deforestationHistoryData.title,
+          image,
+          deforestationHistoryData.description,
+          deforestationHistoryData.layerData
+        );
+        this.deforestationHistories.push(vision);
+      });
+
+      const burningSpotlightsData = this.reportConfig.burningSpotlights;
+      burningSpotlightsData.forEach((burningSpotlightData: Vision) => {
+        const image = burningSpotlightData.image.replace('{bbox}', this.property.bbox);
+        const vision = new Vision(
+          burningSpotlightData.id,
+          burningSpotlightData.title,
+          image,
+          burningSpotlightData.description,
+          burningSpotlightData.layerData
+        );
+        this.burningSpotlights.push(vision);
+      });
+
+      const burnedAreasData = this.reportConfig.burnedAreas;
+      burnedAreasData.forEach((burnedAreaData: Vision) => {
+        const image = burnedAreaData.image.replace('{bbox}', this.property.bbox);
+        const vision = new Vision(
+          burnedAreaData.id,
+          burnedAreaData.title,
+          image,
+          burnedAreaData.description,
+          burnedAreaData.layerData
+        );
+        this.burnedAreas.push(vision);
+      });
+
+      const landsatHistoriesData = this.reportConfig.landsatHistories;
+      landsatHistoriesData.forEach((landsatHistoryData: Vision) => {
+        const image = landsatHistoryData.image.replace('{bbox}', this.property.bbox);
+        const vision = new Vision(
+          landsatHistoryData.id,
+          landsatHistoryData.title,
+          image,
+          landsatHistoryData.description,
+          landsatHistoryData.layerData
+        );
+        this.landsatHistories.push(vision);
+      });
     });
   }
-
 }
